[squash] Rename file
