+++ lib/UtapiClient.js
@@ -1,6 +1,6 @@
 const assert = require('assert');
 const { Logger } = require('werelogs');
-const { methods } = require('../utils/genericOperations');
+const { methods } = require('../utils/S3operations');
 const Datastore = require('./Datastore');
 const { generateKey, generateCounter, generateStateKey } = require('./schema');
 const { errors } = require('arsenal');
