+++ lib/schema.js
@@ -1,5 +1,5 @@
 const { genericOperations, getS3Operation } =
-    require('../utils/genericOperations');
+    require('../utils/S3operations');
 
 // metric type schema
 const stateKeys = {
