+++ tests/functional/testReplay.js
@@ -5,7 +5,7 @@ const UtapiReplay = require('../../lib/UtapiReplay');
 const UtapiClient = require('../../lib/UtapiClient');
 const Datastore = require('../../lib/Datastore');
 const redisClient = require('../../utils/redisClient');
-const { genericOperations } = require('../../utils/genericOperations');
+const { genericOperations } = require('../../utils/S3operations');
 const { getAllResourceTypeKeys } = require('../testUtils');
 const safeJsonParse = require('../../utils/safeJsonParse');
 const localCache = redisClient({
